requirements.txt
apache-airflow==2.8.1
pandas
requests
sqlalchemy
python-dotenv
